moving render/detailed/parents to it's own file
